20230222
-made config.py suitable for local:host, Server as well as new&old router
- started to redesign weatherinfo modul
